New Version of Scraper.py
This is in reply to the issue raised by me. Actually in the previous code, you were sending the driver.page_source to the soup. So in some machines, it may print the file even before the whole page is loaded. So actually it prints the landing page.( I don't know the actual reason, but I guess it would be something to do with the speed of the machine or the parallel processing in it, which sends the file to print even before it is loaded).

So here, I have explicitly forced to wait the driver until it finds the element with the tag 'html' ( :P ), i.e., now the page is fully loaded. So now we can safely pass the driver to the soup and it will  print the destination page.

This helps in going free from the shackles of the machines and OS-es. You can merge this to the master in your repo..

P.S: I was using Ubuntu 14.04 in a 2.1 GHz processor( if  this info is required!)